salesevents/requirements.txt
# sudo pip install -r requirements.txt
django
django-mptt
